Migrate second fetch to utils
