+++ src/components/ListingsContainer/ListingsContainer.js
@@ -2,6 +2,7 @@ import React from 'react';
 import ListingCard from '../ListingCard/ListingCard';
 import Nav from '../Nav/Nav';
 import './ListingsContainer.scss';
+import { fetchListingData } from '../../utils';
 
 class ListingsContainer extends React.Component {
   constructor({ match, location }) {
@@ -14,29 +15,12 @@ class ListingsContainer extends React.Component {
   }
 
   componentDidMount() {
-    this.fetchListingData();
+    fetchListingData(this.areaID).then(listings => this.setState({listings}));
     this.fetchSavedFavorites();
   }
 
   componentDidUpdate() {
     this.updateSavedFavorites();
-  }
-
-  fetchListingData = () => {
-    fetch('http://localhost:3001/api/v1/areas/' + this.areaID)
-     .then(res => res.json())
-     .then(data => {
-      const promises = data.listings.map(listing => {
-        return fetch('http://localhost:3001'+ listing)
-              .then(res => res.json())
-              .then(info => {
-                return {name: info.name,
-                  listingID: info.listing_id
-                  };
-              })
-      })
-      Promise.all(promises).then(listings => this.setState({listings}));
-    })
   }
 
   fetchSavedFavorites = () => {
